#defcfunc戻り値 + DLL戻り値 + hsp3cl int64対応 (Phase 6)
#defcfunc int64戻り値:
- cmdfunc_return_setval: HSPVAR_FLAG_INT64 → stat に int64値をセット
- reffunc_cfunc: funcres == HSPVAR_FLAG_INT64 時に stat から返却

DLL呼び出し戻り値:
- code_expand_and_call/code_expand_next: 戻り値を int → int64_t に変更
- exec_dllcmd: result を int64_t に（64bit環境でraxの64bit値を保持）

hsp3cl (コンソール版):
- v143 + /source-charset:utf-8 に更新
- hspvar_int64.cpp, filepack.cpp, hsp3crypt.cpp, hsp3utfcnv.cpp を追加

hspcmp:
- GetParameterTypeCG/FuncTypeCG/StructTypeCG: int64, float パラメータ追加
- PutStructParam: MPTYPE_INUM64 サイズ(8byte)対応
- MPTYPE_INUM64 = HSPVAR_FLAG_INT64 に修正

テスト: 全機能の包括テスト合格

+++ hsp3/win32gui/hsp3extlib.cpp
@@ -531,9 +531,9 @@ static char *prepare_localstr( char *src, int mode )
 	return dst;
 }
 
-static int code_expand_next( char *, const STRUCTDAT *, int );
+static int64_t code_expand_next( char *, const STRUCTDAT *, int );
 
-int code_expand_and_call( const STRUCTDAT *st )
+int64_t code_expand_and_call( const STRUCTDAT *st )
 {
 	//	パラメータの取得および関数呼び出し（再帰処理による）
 	//
@@ -543,7 +543,7 @@ int code_expand_and_call( const STRUCTDAT *st )
 	//	DLL 関数呼び出し時は st->proc に関数アドレスをセットして
 	//	おかなければなりません（ BindFUNC() により）。
 	//
-	int result;
+	int64_t result;
 
 #ifdef HSP64
 	char *prmbuf = sbAlloc(st->prmmax * sizeof(INT_PTR));
@@ -562,11 +562,11 @@ int code_expand_and_call( const STRUCTDAT *st )
 	return result;
 }
 
-static int code_expand_next( char *prmbuf, const STRUCTDAT *st, int index )
+static int64_t code_expand_next( char *prmbuf, const STRUCTDAT *st, int index )
 {
 	//	次のパラメータを取得（および関数呼び出し）（再帰処理）
 	//
-	int result;
+	int64_t result;
 	HSPAPICHAR *hactmp1 = 0;
 	if ( index == st->prmmax ) {
 		// 関数（またはメソッド）の呼び出し
@@ -715,7 +715,7 @@ int exec_dllcmd( int cmd, int mask )
 {
 	STRUCTDAT *st;
 	FARPROC pFn;
-	int result;
+	int64_t result;
 
 	code_next();							// 次のコードを取得(最初に必ず必要です)
